[CHERRY-PICK] Adding '823d5db40eab964bdcd07412acd9c6b10edcc4b5' to 'CHERRY_PICK_main_2024-04-05_16.27.21' using the python script version 'v1.0.8'.

+++ PresentationLayer/ACADEMIC/SessionMaster.aspx.cs
@@ -396,7 +396,7 @@ public partial class Academic_SessionCreate : System.Web.UI.Page
 
                     //below code ADDED BY SHAILENDRA K ON DATED 13.03.2024 AS PER T-55537 & 55470
 
-                    int sessionRegdCount = Convert.ToInt16(objCommon.LookUp("ACD_STUDENT_RESULT", "COUNT(1) CNT", @"ISNULL(CANCEL,0)=0 AND
+                    int sessionRegdCount = Convert.ToInt32(objCommon.LookUp("ACD_STUDENT_RESULT", "COUNT(1) CNT", @"ISNULL(CANCEL,0)=0 AND
                                                                     SESSIONNO IN (
                                                                     SELECT DISTINCT SM.SESSIONNO FROM ACD_SESSION S 
                                                                     INNER JOIN ACD_SESSION_MASTER SM ON S.SESSIONID=SM.SESSIONID
